+++ R/RcppExports.R
@@ -61,6 +61,10 @@ nalleles <- function(crosstype) {
     .Call('qtl2geno_get_x_covar', PACKAGE = 'qtl2geno', crosstype, is_female, cross_info)
 }
 
+.calc_errorlod <- function(crosstype, probs, genotypes, founder_geno, is_X_chr, is_female, cross_info) {
+    .Call('qtl2geno_calc_errorlod', PACKAGE = 'qtl2geno', crosstype, probs, genotypes, founder_geno, is_X_chr, is_female, cross_info)
+}
+
 .calc_genoprob <- function(crosstype, genotypes, founder_geno, is_X_chr, is_female, cross_info, rec_frac, marker_index, error_prob) {
     .Call('qtl2geno_calc_genoprob', PACKAGE = 'qtl2geno', crosstype, genotypes, founder_geno, is_X_chr, is_female, cross_info, rec_frac, marker_index, error_prob)
 }
@@ -83,6 +87,10 @@ nalleles <- function(crosstype) {
 
 addlog <- function(a, b) {
     .Call('qtl2geno_addlog', PACKAGE = 'qtl2geno', a, b)
+}
+
+subtrlog <- function(a, b) {
+    .Call('qtl2geno_subtrlog', PACKAGE = 'qtl2geno', a, b)
 }
 
 .viterbi <- function(crosstype, genotypes, founder_geno, is_X_chr, is_female, cross_info, rec_frac, marker_index, error_prob) {